Use size_t for SDK_ARRAYSIZE instead
sizeof() returns a size_t.

+++ r5dev/tier0/basetypes.h
@@ -135,7 +135,7 @@
 #define MAX_MAP_NAME 64
 
 #define SDK_VERSION "VGameSDK001" // Increment this with every /breaking/ SDK change (i.e. security/backend changes breaking compatibility).
-#define SDK_ARRAYSIZE(arr) ((int)(sizeof(arr) / sizeof(*arr))) // Name due to IMGUI implementation and NT implementation that we shouldn't share across everywhere.
+#define SDK_ARRAYSIZE(arr) ((sizeof(arr) / sizeof(*arr))) // Name due to IMGUI implementation and NT implementation that we shouldn't share across everywhere.
 
 #ifndef DEDICATED
 #define SDK_DEFAULT_CFG "platform\\cfg\\startup_default.cfg"
